Small fixes after video.

+++ create_dataset.py
@@ -7,9 +7,9 @@ import numpy as np
 import tensorflow as tf
 
 def _int64_feature(value):
-  return tf.train.Feature(int64_list=tf.train.Int64List(value=[value]))
+    return tf.train.Feature(int64_list=tf.train.Int64List(value=[value]))
 def _bytes_feature(value):
-  return tf.train.Feature(bytes_list=tf.train.BytesList(value=[value]))
+    return tf.train.Feature(bytes_list=tf.train.BytesList(value=[value]))
 
 def load_image(addr):
     # read an image and resize to (224, 224)
@@ -17,10 +17,8 @@ def load_image(addr):
     img = cv2.imread(addr)
     if img is None:
         return None
-
     img = cv2.resize(img, (224, 224), interpolation=cv2.INTER_CUBIC)
     img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
-    img = img.astype(np.float32)
     return img
  
 def createDataRecord(out_filename, addrs, labels):
@@ -63,7 +61,7 @@ c = list(zip(addrs, labels))
 shuffle(c)
 addrs, labels = zip(*c)
     
-# Divide the hata into 60% train, 20% validation, and 20% test
+# Divide the data into 60% train, 20% validation, and 20% test
 train_addrs = addrs[0:int(0.6*len(addrs))]
 train_labels = labels[0:int(0.6*len(labels))]
 val_addrs = addrs[int(0.6*len(addrs)):int(0.8*len(addrs))]
